+++ 11-react/L05-props/02-prop-types/src/lib/components/Aside.js
@@ -22,7 +22,7 @@ const  Aside = ({ links, titulo }) => {
   }
 
   const children = links.map(props => {
-    return <AsideLink {...props} />
+    return <AsideLink key={props.texto} {...props} />
   })
   
   return (
