+++ 11-react/L05-props/02-prop-types/src/lib/components/Page.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Header from './header'
-import MainSection from './main-section'
-import Aside from './aside'
+import Header from './Header'
+import MainSection from './MainSection'
+import Aside from './Aside'
 
 const Page = ({
   headerTitulo,
