+++ 11-react/L05-props/04-children/src/lib/components/MainSection.js
@@ -24,7 +24,7 @@ const MainSection = ({ titulo, articles }) => {
   }
   
   const children = articles.map(props => {
-    return <Article {...props} />
+    return <Article key={props.titulo} {...props} />
   })
 
   return (
