+++ 11-react/L05-props/04-children/webpack.config.js
@@ -13,7 +13,7 @@ const config = {
     'react-hot-loader/patch',
     'webpack-dev-server/client?http://localhost:8080',
     'webpack/hot/only-dev-server',
-    './main.js'
+    './index.js'
   ],
 
   devServer: {
