+++ 11-react/L05-props/04-children/webpack.prod.config.js
@@ -6,7 +6,7 @@ const config = {
   devtool: 'cheap-module-source-map',
 
   entry: [
-    './main.js'
+    './index.js'
   ],
 
   context: resolve(__dirname, 'src'),
